requirements.txt
numpy==1.22.3
pillow==9.0.1
Flask==2.0.3
mysql-connector==2.2.9